feature: renamed package to conform to standards
