+++ internal/cmd/beta/security-group/security_group.go
@@ -1,9 +1,14 @@
-package securitygroup
+package security_group
 
 import (
-	"github.com/stackitcloud/stackit-cli/internal/cmd/beta/security-group/rule"
+	"github.com/stackitcloud/stackit-cli/internal/cmd/beta/security-group/create"
+	"github.com/stackitcloud/stackit-cli/internal/cmd/beta/security-group/delete"
+	"github.com/stackitcloud/stackit-cli/internal/cmd/beta/security-group/describe"
+	"github.com/stackitcloud/stackit-cli/internal/cmd/beta/security-group/list"
+	"github.com/stackitcloud/stackit-cli/internal/cmd/beta/security-group/update"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/args"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/print"
+
 	"github.com/stackitcloud/stackit-cli/internal/pkg/utils"
 
 	"github.com/spf13/cobra"
@@ -12,8 +17,8 @@ import (
 func NewCmd(p *print.Printer) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "security-group",
-		Short: "Provides functionality for security groups",
-		Long:  "Provides functionality for security groups.",
+		Short: "manage security groups.",
+		Long:  "manage the lifecycle of security groups.",
 		Args:  args.NoArgs,
 		Run:   utils.CmdHelp,
 	}
@@ -22,5 +27,11 @@ func NewCmd(p *print.Printer) *cobra.Command {
 }
 
 func addSubcommands(cmd *cobra.Command, p *print.Printer) {
-	cmd.AddCommand(rule.NewCmd(p))
+	cmd.AddCommand(
+		create.NewCmd(p),
+		delete.NewCmd(p),
+		describe.NewCmd(p),
+		list.NewCmd(p),
+		update.NewCmd(p),
+	)
 }
